# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: STWV Multi-Company Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a contract
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contracts/{id}/send:
    post:
      operationId: sendContract
      tags: [contracts]
      summary: Send contract to client via email with PDF attached
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendDocumentResponse"
        "400":
          description: No client email configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Email delivery failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contracts/{id}/pdf:
    get:
      operationId: getContractPdf
      tags: [contracts]
      summary: Download contract as PDF
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contract-templates:
    get:
      operationId: listContractTemplates
      tags: [contracts]
      summary: List contract templates
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of contract templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractTemplate"
    post:
      operationId: createContractTemplate
      tags: [contracts]
      summary: Create a contract template
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractTemplateBody"
      responses:
        "201":
          description: Created template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractTemplate"
  /contract-templates/{id}:
    patch:
      operationId: updateContractTemplate
      tags: [contracts]
      summary: Update a contract template
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractTemplateBody"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractTemplate"

  # ───── INVOICES ─────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    SendDocumentResponse:
      type: object
      properties:
        success:
          type: boolean
        email:
          type: string
      required:
        - success
        - email
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Company ───
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        legalForm:
          type: string
        country:
          type: string
        taxRegime:
          type: string
          description: "One of: vat, gst, none"
        taxNumber:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        bankDetails:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        currency:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - legalForm
        - country
        - taxRegime
        - currency
        - isActive
        - createdAt
        - updatedAt
...
    UserBasic:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        role:
          type: string
      required:
        - id
        - email
        - role
...
    ProjectBasic:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
        - id
        - name

    # ─── Contracts ───
    Contract:
      type: object
      properties:
        id:
          type: integer
        contractNumber:
          type: string
        type:
          type: string
          description: "One of: client_service, freelancer_service"
        companyId:
          type: integer
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        offerId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
        status:
          type: string
          description: "One of: draft, sent, signed, cancelled"
        signedAt:
          type: ["string", "null"]
          format: date-time
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        company:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Company"
        client:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/UserBasic"
        project:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ProjectBasic"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - contractNumber
        - type
        - companyId
        - title
        - content
        - status
        - createdAt
        - updatedAt
    CreateContractBody:
      type: object
      properties:
        type:
          type: string
          description: "One of: client_service, freelancer_service"
        companyId:
          type: integer
        projectId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        title:
          type: string
        content:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
      required:
        - type
        - companyId
        - title
        - content
    UpdateContractBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          type: string
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
    ContractTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        content:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - content
        - isDefault
        - createdAt
        - updatedAt
    CreateContractTemplateBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        content:
          type: string
        isDefault:
          type: boolean
      required:
        - name
        - type
        - content
    UpdateContractTemplateBody:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        isDefault:
          type: boolean

    # ─── Messages ───
...